KTOR-2763 Fix Warnings and Enable -Werror flag (#2496)
* KTOR-2763 Fix warnings in ktor-io

* KTOR-2763 Fix warnings in ktor-utils

* KTOR-2763 Fix warnings in ktor-http

* KTOR-2763 Fix warnings in ktor-network

* KTOR-2763 Fix warnings in ktor-features

* KTOR-2763 Fix warnings in ktor-server

* KTOR-2763 Fix warnings in ktor-client

* KTOR-2763 Fix warnings in ktor-shared & ktor-test-dispatcher

* KTOR-2763 Update configuration

+++ ktor-server/ktor-server-plugins/ktor-server-websockets/jvm/test/io/ktor/tests/websocket/ParserTest.kt
@@ -78,7 +78,7 @@ class ParserTest {
         assertEquals(1, parser.length)
         assertTrue { parser.bodyReady }
 
-        assertEquals('1', buffer.get().toChar())
+        assertEquals('1', buffer.get().toInt().toChar())
         parser.bodyComplete()
 
         parser.frame(buffer)
@@ -89,7 +89,7 @@ class ParserTest {
         assertEquals(1, parser.length)
         assertTrue { parser.bodyReady }
 
-        assertEquals('2', buffer.get().toChar())
+        assertEquals('2', buffer.get().toInt().toChar())
         parser.bodyComplete()
 
         parser.frame(buffer)
@@ -100,7 +100,7 @@ class ParserTest {
         assertEquals(1, parser.length)
         assertTrue { parser.bodyReady }
 
-        assertEquals('3', buffer.get().toChar())
+        assertEquals('3', buffer.get().toInt().toChar())
         parser.bodyComplete()
 
         assertFalse { buffer.hasRemaining() }
